requirements.txt
asgiref==3.7.2
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
distlib==0.3.7
Django==4.2.7
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
dulwich==0.21.6
fastjsonschema==2.19.0
filelock==3.13.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.3.0
MarkupSafe==2.1.3
more-itertools==10.1.0
msgpack==1.0.7
openapi-codec==1.3.2
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2==2.9.9
ptyprocess==0.7.0
pyproject_hooks==1.0.0
pytest==7.4.3
pytest-django==4.7.0
python-decouple==3.8
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==69.0.2
shellingham==1.5.4
simplejson==3.19.2
sqlparse==0.4.4
tomlkit==0.12.3
trove-classifiers==2023.11.22
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.7
